+++ CHANGES.md
@@ -2,6 +2,8 @@
 
 ### Fixes
 
+* xcube server Python scripts can now import modules from
+  the script's directory.
 * xcube serve correctly resolves relative paths to datasets (#758)
 
 ## Changes in 0.13.0.dev3